refactor: extract Command code since circ dep
The AuthCommandBackend class is a higher level
class and shouldn't be with the utils functions
since it causes circular dep issues.

It has been extracted in to its own file in
src/auth/commands.ts
